+++ src/core/gateways/file-storage-gateway.ts
@@ -1,4 +1,0 @@
-export type {
-    StoredObject,
-    FileStorageGateway,
-} from "../../gateways/files/gateway.js";